+++ ntex/src/client/error.rs
@@ -104,7 +104,7 @@ pub enum ConnectError {
     /// SSL error
     #[cfg(feature = "openssl")]
     #[error("{0}")]
-    SslError(#[source] Rc<SslError>),
+    SslError(#[source] Rc<dyn StdError>),
 
     /// SSL Handshake error
     #[cfg(feature = "openssl")]
@@ -344,3 +344,7 @@ impl From<ClientBuilderError> for io::Error {
         io::Error::other(err)
     }
 }
+
+#[doc(hidden)]
+#[deprecated(since = "3.7.0", note = "Use ntex::client::error::ClientError instead")]
+pub type SendRequestError = ClientError;