+++ ntex/src/client/h2proto.rs
@@ -2,29 +2,41 @@ use std::{fmt::Write, future::poll_fn, io, rc::Rc};
 
 use ntex_h2::{self as h2, client::RecvStream, client::SimpleClient, frame};
 
-use crate::error::Error;
-use crate::http::ResponseHead;
+use crate::error::{Error, ErrorMapping, with_service};
 use crate::http::body::{Body, BodySize, MessageBody};
 use crate::http::header::{self, HeaderMap, HeaderValue};
-use crate::http::{Method, Payload, Version, h2::Payload as H2Payload};
+use crate::http::{Method, Payload, ResponseHead, Version, h2::Payload as H2Payload};
 use crate::time::{Millis, timeout_checked};
 use crate::util::{ByteString, Bytes, BytesMut, Either, select};
 
-use super::ClientRawRequest;
-use super::error::{ClientError, ConnectError};
+use super::{ClientRawRequest, error::ClientError, error::ConnectError};
 
 pub(super) async fn send_request(
     client: H2Client,
     req: ClientRawRequest,
     body: Body,
     timeout: Millis,
-) -> Result<(ResponseHead, Payload), ClientError> {
+) -> Result<(ResponseHead, Payload), Error<ClientError>> {
+    with_service(
+        client.service(),
+        send_request_inner(client, req, body, timeout),
+    )
+    .await
+}
+
+async fn send_request_inner(
+    client: H2Client,
+    req: ClientRawRequest,
+    body: Body,
+    timeout: Millis,
+) -> Result<(ResponseHead, Payload), Error<ClientError>> {
     #[cfg(feature = "trace")]
     log::trace!(
         "{}: Sending client request: {req:?} {:?}",
         client.client.tag(),
         body.size()
     );
+
     let length = body.size();
     let eof = if req.head.method == Method::HEAD {
         true
@@ -81,7 +93,7 @@ pub(super) async fn send_request(
         .client
         .send(req.head.method.clone(), path, hdrs, eof)
         .await
-        .map_err(Error::into_error)?;
+        .into_error()?;
 
     // send body
     if !eof {
@@ -97,13 +109,13 @@ pub(super) async fn send_request(
 
     timeout_checked(timeout, get_response(rcv_stream))
         .await
-        .map_err(|()| ClientError::Timeout)
+        .map_err(|()| Error::from(ClientError::Timeout))
         .and_then(|res| res)
 }
 
 async fn get_response(
     rcv_stream: RecvStream,
-) -> Result<(ResponseHead, Payload), ClientError> {
+) -> Result<(ResponseHead, Payload), Error<ClientError>> {
     let h2::Message { stream, kind } = rcv_stream
         .recv()
         .await
@@ -216,23 +228,25 @@ async fn get_response(
                     };
                     Ok((head, payload))
                 }
-                None => Err(ClientError::H2(h2::OperationError::Connection(
-                    h2::ConnectionError::MissingPseudo("Status"),
+                None => Err(Error::from(ClientError::H2(
+                    h2::OperationError::Connection(h2::ConnectionError::MissingPseudo(
+                        "Status",
+                    )),
                 ))),
             }
         }
-        h2::MessageKind::Disconnect(err) => Err(ClientError::H2(err.into_error())),
-        _ => Err(ClientError::Error(Rc::new(io::Error::new(
+        h2::MessageKind::Disconnect(err) => Err(err.map(ClientError::H2)),
+        _ => Err(Error::from(ClientError::Error(Rc::new(io::Error::new(
             io::ErrorKind::Unsupported,
             "Unexpected h2 message",
-        )))),
+        ))))),
     }
 }
 
 async fn send_body(
     mut body: Body,
     stream: &h2::client::SendStream,
-) -> Result<(), ClientError> {
+) -> Result<(), Error<ClientError>> {
     loop {
         match poll_fn(|cx| body.poll_next_chunk(cx)).await {
             Some(Ok(b)) => {
@@ -243,20 +257,13 @@ async fn send_body(
                     stream.id(),
                     b.len()
                 );
-                stream
-                    .send_payload(b, false)
-                    .await
-                    .map_err(Error::into_error)?;
+                stream.send_payload(b, false).await.into_error()?;
             }
-            Some(Err(e)) => return Err(e.into()),
+            Some(Err(e)) => return Err(Error::from(ClientError::from(e))),
             None => {
                 #[cfg(feature = "trace")]
                 log::trace!("{}: {:?} eof of send stream ", stream.tag(), stream.id());
-                stream
-                    .send_payload(Bytes::new(), true)
-                    .await
-                    .map_err(Error::into_error)?;
-                return Ok(());
+                return stream.send_payload(Bytes::new(), true).await.into_error();
             }
         }
     }
@@ -274,6 +281,10 @@ impl H2Client {
 
     pub(super) fn tag(&self) -> &'static str {
         self.client.tag()
+    }
+
+    pub(super) fn service(&self) -> &'static str {
+        self.client.service()
     }
 
     pub(super) fn close(&self) {